oss_backend/requirements.txt
Django==5.2.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
drf-yasg==1.21.7
psycopg2-binary
python-dotenv==1.0.0
gunicorn==23.0.0
django-filter==23.5
django-cors-headers==4.3.1
django-prometheus==2.3.1
prometheus-client==0.19.0
django-mptt==0.15.0
requests==2.32.0
django-mailgun-provider==0.2.3
pymongo==4.13.0
redis==6.1.0
setuptools==80.7.1
whitenoise==6.6.0
